Move component interface from shared shipping types file
